HeaderIDE changes for timer
