+++ src/containers/IDEContainer.jsx
@@ -1,7 +1,11 @@
 import React from "react";
 
 function IDEContainer() {
-  return <div>IDEContainer</div>;
+  return (
+    <div>
+      IDEContainer
+    </div>
+  );
 }
 
 export default IDEContainer;
